Clean out API based on new discussion
